UPSTREAM: crypto: arm/aes - streamline AES-192 code path
This trims off a couple of instructions of the total size of the
core AES transform by reordering the final branch in the AES-192
code path with the rounds that are performed regardless of whether
the branch is taken or not. Other than the slight size reduction,
this has no performance benefit.

Fix up a comment regarding the prototype of this function while
we're at it.

Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 6499e8cfaa8f5d041b20af24d8409dec9f3ac3d0)
Change-Id: If77f938fba6b3cbfb3af577d40f668f18aef8290
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>
Signed-off-by: MOVZX <movzx@yahoo.com>