pwm: pwm-qpnp: Re-enable PWM output after changing mode
commit c8d3578 ("pwm: pwm-qpnp: Ignore changing period setting if
it's not requested") had fixed an issue to avoid glitches seen in PWM
output when the PWM period is not changed and the duty cycle alone is
changed. However, that didn't fix the issue completely when the LPG
channel operation mode might be configured differently even when the
duty cycle is changed and the period is unchanged. This will lead to LPG
channel getting configured to operate in a wrong mode. Fix this by
enabling PWM output whenever there is a change in the mode for LPG
channel requested by the client.

CRs-Fixed: 2070945
Change-Id: Ida6145a7cb9016e9e9116fbcd6b8ba3f6c5b3bc1
Signed-off-by: Fenglin Wu <fenglinw@codeaurora.org>
Signed-off-by: MOVZX <movzx@yahoo.com>