drm/msm/sde: update pipes on dirty plane properties
Normal 'has mode changed' checks are sufficient for determining
whether the entire pipe needs to be reprogrammed, but partial
updating of the pipe configuration may also be required due to
user property updates. This change modifies the mode set logic
to also consider incoming property sets that have occurred
since the last update to determine what needs to be reconfigured.

Change-Id: Ib2d15c3882bd4bdb6bc6b53c7f160e6da847a2ff
Signed-off-by: Clarence Ip <cip@codeaurora.org>
Signed-off-by: MOVZX <movzx@yahoo.com>